experiment(transpiler): Transpile service definitions to Express.js app
A microservice is specified with a TypeScript class declaration.
The service is identified by the class name and must be universally
unique. Handlers on a service are specified with method declarations.
Handlers receive a typed object as input, and return a typed object
as output.

This change enables a barebone (NoOp input) and
standalone (no dependency) service. Service definitions
"compile" to Express.js applications, with requests handlers
for each method. Input values are not being handled. Static
response return statements are transformed into Express.js
response statements.